This was the first provided example to scan a core and stitch everything with a few key additions:
Note: 'Core Sample' must be clicked upon sample add to get make this work. The software is not ready to merge to main yet.

1. Golden section search autofocus (~8 seconds to get autofocus)
2. Autofocusing every other image instead of every image (does not need autofocus on each new frame)
3. Adding cores to be digitized by navigating the camera to the bottom of the core instead of the midpoint
4. Intelligent stop-capture when the camera moves past the top most point of the core

This has not accounted for how this changes the acquisition of metadata and the assignment to the metadata.json file. This also needs to be integrated into cookie capturing too.